fixing #40
create_analysis_prop_select_one will now shows the number of missing values instead of the empty line.
create_analysis_prop_select_multiple will add a line with the analysis_var_value NA and the number of NA if there is any.

For the numerical analysis (mean, median, ratio), the n will be the number of rows used in the calculation, it won't add a line to show the number of missing values