fixing create_analysis_median
the function was calculating the mean instead of the median when a subset of the dataframe had only NAs. All values were the mean, instead of the median.

error was coming from the tryCatch that was not catching the correct error. The tryCatch was initially introduced to have the function run when only NA. It was now catching another error, i.e. when only subset has only missing value.